Consolidate package structure
