+++ org.moflon.tie.gt.ide/src/org/moflon/tie/gt/ide/core/patterns/VariableLookup.java
@@ -16,7 +16,7 @@ import org.moflon.tie.gt.compiler.democles.CodeConventions;
 import org.moflon.tie.gt.compiler.democles.pattern.DemoclesPatternType;
 import org.moflon.tie.gt.controlflow.democles.CFVariable;
 import org.moflon.tie.gt.ide.core.patterns.util.Patterns;
-import org.moflon.tie.gt.ide.core.runtime.utilities.TypeLookup;
+import org.moflon.tie.gt.ide.core.patterns.util.TypeLookup;
 
 class VariableLookup {
 	private final Map<DemoclesPatternType, Map<Object, EMFVariable>> data = new HashMap<>();
