+++ org.moflon.tie.gt.ide/src/org/moflon/tie/gt/ide/core/patterns/util/TypeLookup.java
@@ -1,4 +1,4 @@
-package org.moflon.tie.gt.ide.core.runtime.utilities;
+package org.moflon.tie.gt.ide.core.patterns.util;
 
 import java.util.Collection;
 import java.util.List;
